gps_pi/requirements.txt
requests>=2.31.0
paho-mqtt>=1.6.1
pyserial>=3.5
pynmea2>=1.19.0
Pillow>=9.0.0
adafruit-circuitpython-ssd1306
Adafruit-Blinka
